backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# AI/ML - Groq and Gemini
groq>=0.4.0
google-generativeai>=0.3.0

# Document processing
pypdf>=3.0.0
pdfminer.six>=20231105
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0